Better distinguish links to issue tracking vs tracker
Our documentation page about how we track issues (issue tracking) was
often referred to as the issue tracker and in several places used an
external link style. That made it hard to know if the link would go to
the actual issue tracker or to the site page that describes our issue
tracking process.

Fix the links to have consistent format and naming and add links to the
actual issue tracker or site page as appropriate. Also add a link to the
contributing page from the bug-fixing page since those are closely
related.

Change-Id: I16e7df3f5e4403706732a7ccab11009fcab75ab7