Move Change 422518 into bug-fixes of Gerrit v3.11 release notes
Change 422518 introduced the nanoseconds tracking of performance values
only for the purpose of avoiding reporting 0ms as execution times.
The unit remains expressed in ms and the nanoseconds tracking
is used to produce a more accurate value with decimals, rather than
reporting 0ms.

The existing applications and tools should continue to work as before
and this improved accurancy is not breaking any existing change.

Change-Id: Ie6a6c81efe9082f5e4b12fc3d0483cdb49dabb7b