Add references to Issue 309098227 on v3.9, v3.10 and v3.11
The Issue 309098227 was detected in v3.8 but never addressed
therefore it is also impacting v3.9, v3.10 and v3.11.

Differently from all the other issues, the consequences of not
implementing the workaround can be severe, including the repository
corruption, therefore it is worth mentioning it in the release
notes in the "Known Issues" section.

Change-Id: I9c9a920ccd51d48354e106d0bc70f3ef92be88eb