Support Jujutsu: Update design after discussion with Git team
It seems the Git team is willing to preserve the 'change-id' commit
header on rebase and cherry-pick. To align with this the Jujutsu team
intends to add an option to their `duplicate` command that lets users
pick if they want a new change ID or the old one. This would solve the
compatibility issues between Jujutsu and Git users and also makes
tracking cherry-picks in Gerrit explicitly unnecessary. Update the
design doc accordingly.

Change-Id: I2d63607318ac44732d5c4fbaa95f0831d1422f7a
Signed-off-by: Edwin Kempin <ekempin@google.com>