Gerrit v3.13.0 release notes: remove Change 458724 from breaking changes
Change 458724 is not worth being mentioned standalone and flagged as a
breaking change. It is already mentioned in the highlights and is not
even breaking anything because of the deprecation of the HTTP passwords
would not impact any existing functionality.

Also, remove the mention of Change 458727 as the deprecation is already
discussed in the highlights section.

Change-Id: I084bc334d4b125710b26595a9c580cc5539b8f34